expense-agent/requirements.txt
slack-bolt>=1.18.0
slack-sdk>=3.25.0
google-api-python-client>=2.100.0
google-auth-httplib2>=0.2.0
google-auth-oauthlib>=1.1.0
anthropic>=0.34.0
Pillow>=10.0.0
pillow-heif>=0.13.0
python-dotenv>=1.0.0
pytest>=7.0.0
